registry: move search-related code to separate files
